@@ -13,40 +13,41 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.2",
-    "@angular/animations": "5.1.3",
-    "@angular/cdk": "5.0.4",
-    "@angular/common": "5.1.3",
-    "@angular/compiler": "5.1.3",
-    "@angular/core": "5.1.3",
-    "@angular/forms": "5.1.3",
-    "@angular/http": "5.1.3",
-    "@angular/material": "5.0.4",
-    "@angular/platform-browser": "5.1.3",
-    "@angular/platform-browser-dynamic": "5.1.3",
-    "@angular/router": "5.1.3",
+    "@angular/animations": "5.2.7",
+    "@angular/cdk": "5.2.3",
+    "@angular/common": "5.2.7",
+    "@angular/compiler": "5.2.7",
+    "@angular/core": "5.2.7",
+    "@angular/forms": "5.2.7",
+    "@angular/http": "5.2.7",
+    "@angular/material": "5.2.3",
+    "@angular/platform-browser": "5.2.7",
+    "@angular/platform-browser-dynamic": "5.2.7",
+    "@angular/router": "5.2.7",
     "core-js": "2.4.1",
-    "rxjs": "5.5.2",
-    "zone.js": "0.8.14"
+    "rxjs": "5.5.6",
+    "zone.js": "0.8.19",
+    "@types/file-saver": "1.3.0",
+    "file-saver": "1.3.3"
   },
   "devDependencies": {
-    "@angular/cli": "1.6.3",
-    "@angular/compiler-cli": "5.1.3",
-    "@angular/language-service": "5.1.3",
-    "@types/jasmine": "~2.5.53",
+    "@angular/cli": "1.7.2",
+    "@angular/compiler-cli": "5.2.7",
+    "@angular/language-service": "5.2.7",
+    "@types/jasmine": "~2.8.3",
     "@types/jasminewd2": "~2.0.2",
     "@types/node": "~6.0.60",
-    "codelyzer": "4.0.1",
-    "jasmine-core": "~2.6.2",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
-    "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "1.2.1",
+    "codelyzer": "^4.0.1",
+    "jasmine-core": "~2.8.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~2.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "^1.2.1",
     "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "0.2.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "~5.1.2",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.7.0",
-    "typescript": "~2.4.2"
+    "ts-node": "~4.1.0",
+    "tslint": "~5.9.1",
+    "typescript": "~2.5.3"
   }
 }
