@@ -1,26 +1,25 @@
 import { Injectable } from '@angular/core';
 import { HttpClient, HttpHeaders, HttpErrorResponse } from '@angular/common/http';
 
-import { Auth, User } from '../../auth/auth.model';
+import { Auth } from '../../auth/auth.model';
 import { AppConstants } from '../../app-constants';
 import { StaticMethods } from '../../utils/static-methods';
 
 import { Observable } from 'rxjs/Observable';
 import { ErrorObservable } from 'rxjs/observable/errorObservable';
-import { catchError, tap } from 'rxjs/operators';
+import { catchError, tap, share } from 'rxjs/operators';
 
-import 'rxjs/add/operator/toPromise';
 import 'rxjs/add/observable/of';
 
 import { Subject } from 'rxjs/Subject';
 
-import { PersonType } from '../../shared/enums/person-type.enum';
 import { ModalManager } from './modal-manager';
 
 
 @Injectable()
 export class AuthService {
 
+  private isTokenChecked = false;
   private _token: string;
   private _auth: Auth;
   public fcm_token: string;
@@ -31,16 +30,8 @@ export class AuthService {
     return this._token;
   }
 
-  get user() {
-    return this._auth ? this._auth.user : undefined;
-  }
-
-  get unreadNots() {
-    return this._auth ? this._auth.unread_nots : undefined;
-  }
-
-  set unreadNots(unreadNots) {
-    this._auth.unread_nots = unreadNots;
+  get user(): Auth {
+    return this._auth ? this._auth : undefined;
   }
 
   loginSbj = new Subject<any>();
@@ -52,52 +43,68 @@ export class AuthService {
   ) { }
 
   login(user): Observable<Auth> {
-    return this.http.post<Auth>(`${AppConstants.API_ENDPOINT}login`, user)
+    return this.http.post<Auth>(`${AppConstants.API_ENDPOINT}sessions`, user)
       .pipe(
-      catchError((err, caught) => {
-        this.mm.closeLoadingDialog();
-        return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
-      }),
-      tap((res) => {
-        this._token = res.token;
-        this._auth = res;
-        sessionStorage.setItem('token', res.token);
-        this.loginSbj.next(res.user);
-        return res;
-      })
+        catchError((err, caught) => {
+          this.mm.closeLoadingDialog();
+          return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
+        }),
+        tap((res) => {
+          this._token = res.token;
+          this._auth = res;
+          sessionStorage.setItem('token', res.token);
+          this.loginSbj.next(res.user);
+          return res;
+        })
       );
   }
 
   signup(account): Observable<any> {
     return this.http.post(`${AppConstants.API_ENDPOINT}users`, account)
       .pipe(
-      catchError((err, caught) => {
-        this.mm.closeLoadingDialog();
-        return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
-      })
+        catchError((err, caught) => {
+          this.mm.closeLoadingDialog();
+          return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
+        })
       );
   }
 
-  check(): Promise<any> {
+  check(): Observable<any> {
     this._token = sessionStorage.getItem('token');
-    return this.http.get<Auth>(`${AppConstants.API_ENDPOINT}check?fcm_token=${this.fcm_token}`)
-      .toPromise()
-      .then((res) => {
-        this._auth = res;
-        this.loginSbj.next(res.user);
-        return res.user;
-      })
-      .catch((err) => {
-        this._token = undefined;
-        this._auth = undefined;
-        sessionStorage.removeItem('token');
-        return Promise.reject(err);
-      });
+    console.log('user', this.user);
+    console.log('token', this.token);
+    if (this.isTokenChecked) {
+      if (this.user) {
+        return Observable.of(this.user);
+      } else {
+        this.mm.closeLoadingDialog();
+        return ErrorObservable.create('');
+      }
+    }
+
+    return this.http.get<Auth>(`${AppConstants.API_ENDPOINT}sessions/check`)
+      .pipe(
+        share(),
+        tap((res) => {
+          this.isTokenChecked = true;
+          this._token = res.token;
+          this._auth = res;
+          this.loginSbj.next(res);
+          return res;
+        }),
+        catchError((err, caught) => {
+          this.isTokenChecked = true;
+          this._auth = undefined;
+          this._token = undefined;
+          sessionStorage.removeItem('token');
+          this.mm.closeLoadingDialog();
+          return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
+        }));
   }
 
   logout(params?): Promise<any> {
 
-    this.http.delete(`${AppConstants.API_ENDPOINT}logout`);
+    this.http.delete(`${AppConstants.API_ENDPOINT}sessions/logout`);
 
     this._token = undefined;
     this._auth = undefined;
@@ -111,20 +118,20 @@ export class AuthService {
       responseType: 'text',
     })
       .pipe(
-      catchError((err, caught) => {
-        this.mm.closeLoadingDialog();
-        return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
-      })
+        catchError((err, caught) => {
+          this.mm.closeLoadingDialog();
+          return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
+        })
       );
   }
 
   recoverPassword(user) {
     return this.http.post(`${AppConstants.API_ENDPOINT}users/update_password`, user)
       .pipe(
-      catchError((err, caught) => {
-        this.mm.closeLoadingDialog();
-        return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
-      })
+        catchError((err, caught) => {
+          this.mm.closeLoadingDialog();
+          return ErrorObservable.create(StaticMethods.handleHttpResponseError(err));
+        })
       );
   }
 }
