@@ -5,32 +5,15 @@ import {
   RouterStateSnapshot
 } from '@angular/router';
 import { AuthService } from './auth.service';
-
-import { PersonType } from '../../shared/enums/person-type.enum';
-
-const redirect = (user, router, which) => {
-  if (user) {
-    switch (user.type_user) {
-      case PersonType.ABOGADO:
-        router.navigate(['/abogado']);
-        break;
-      case PersonType.ADMIN:
-        router.navigate(['/admin']);
-        break;
-      default:
-        router.navigate(['/usuario']);
-        break;
-    }
-  } else {
-    router.navigate(['']);
-  }
-};
+import { UserType } from '../../shared/models/user.model';
+import { ModalManager } from './modal-manager';
 
 @Injectable()
 export class LoginAuthGuard implements CanActivate {
   constructor(
     private router: Router,
-    private authService: AuthService) { }
+    private authService: AuthService,
+    private mm: ModalManager) { }
 
   canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean {
     const url: string = state.url;
@@ -41,32 +24,20 @@ export class LoginAuthGuard implements CanActivate {
   checkLogin(url: string): boolean {
     if (this.authService.user) { return true; }
 
+    console.log('#LoginAuthGuard');
     this.authService.redirectUrl = url;
 
-    redirect(this.authService.user, this.router, 1);
-
-    return false;
-  }
-}
-
-@Injectable()
-export class LawyerAuthGuard implements CanActivate {
-  constructor(
-    private router: Router,
-    private authService: AuthService) { }
-
-  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean {
-    const url: string = state.url;
-
-    return this.checkLogin(url);
-  }
-
-  checkLogin(url: string): boolean {
-    if (this.authService.user && this.authService.user.type_user === PersonType.ABOGADO) { return true; }
-
-    this.authService.redirectUrl = url;
-
-    redirect(this.authService.user, this.router, 2);
+    this.mm.showLoadingDialog();
+    this.authService.check()
+      .subscribe(res => {
+        console.log('#LoginAuthGuard#succ');
+        this.router.navigate([this.authService.redirectUrl]);
+      }, err => {
+        console.log('#LoginAuthGuard#err');
+        this.router.navigate(['/login']);
+      }, () => {
+        this.mm.closeLoadingDialog();
+      });
 
     return false;
   }
@@ -81,41 +52,15 @@ export class AdminAuthGuard implements CanActivate {
   canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean {
     const url: string = state.url;
 
-    return this.checkLogin(url);
+    return this.isAdmin(url);
   }
 
-  checkLogin(url: string): boolean {
-    if (this.authService.user && this.authService.user.type_user === PersonType.ADMIN) { return true; }
+  isAdmin(url: string): boolean {
+    if (this.authService.user.role === UserType.ADMIN) { return true; }
 
     this.authService.redirectUrl = url;
 
-    redirect(this.authService.user, this.router, 3);
-
-    return false;
-  }
-}
-
-@Injectable()
-export class PersonAuthGuard implements CanActivate {
-  constructor(
-    private router: Router,
-    private authService: AuthService) { }
-
-  canActivate(route: ActivatedRouteSnapshot, state: RouterStateSnapshot): boolean {
-    const url: string = state.url;
-
-    return this.checkLogin(url);
-  }
-
-  checkLogin(url: string): boolean {
-    if (this.authService.user &&
-      (this.authService.user.type_user === PersonType.JURIDICA || this.authService.user.type_user === PersonType.NATURAL)) {
-      return true;
-    }
-
-    this.authService.redirectUrl = url;
-
-    redirect(this.authService.user, this.router, 4);
+    this.router.navigate(['/infracciones']);
 
     return false;
   }
